requirements.txt
attrs==25.3.0
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
h11==0.14.0
idna @ file:///croot/idna_1714398848350/work
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==8.3.5
python-dotenv==1.0.1
requests @ file:///croot/requests_1730999120400/work
selenium==4.17.2
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve @ file:///work/perseverance-python-buildout/croot/soupsieve_1698866207280/work
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3 @ file:///croot/urllib3_1737133630106/work
webdriver-manager==4.0.2
wheel==0.45.1
wsproto==1.2.0
